requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.88.0
filelock==3.8.2
gensim==3.8.3
h11==0.14.0
huggingface-hub==0.11.1
idna==3.4
joblib==1.2.0
navec==0.10.0
nltk==3.8
numpy==1.23.5
packaging==22.0
Pillow==9.3.0
pydantic==1.10.2
PyYAML==6.0
razdel==0.5.0
regex==2022.10.31
requests==2.28.1
scikit-learn==1.2.0
scipy==1.9.3
sentence-transformers==2.2.2
sentencepiece==0.1.97
smart-open==6.3.0
sniffio==1.3.0
starlette==0.22.0
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1+cpu
torchvision==0.14.1+rocm5.2
tqdm==4.64.1
transformers==4.25.1
typing-extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
aioredis==1.3.1
pycryptodome==3.11.0